algod1: widen network edge offsets to uint64_t
ampinfo_s::link_start and Network_state::count are offsets into / the
running total of network_v, the d=1 edge list. As unsigned int they
overflow past UINT_MAX on a large enough dataset (total edges can reach
sequence_count * (7+4)), corrupting the link_start offsets and the
subsequent network_v indexing. Widen both, and the write_network_file
count parameter, to uint64_t (network_v's natural index type). The
element type and the per-amplicon link_count stay unsigned int.

+++ src/utils/algod1_internal.h
@@ -56,8 +56,8 @@ struct ampinfo_s
   unsigned int generation {0U};
   unsigned int next {no_swarm};        /* amp id of next amplicon in swarm */
   unsigned int graft_cand {no_swarm};  /* amp id of potential grafting parent (fastid.) */
-  unsigned int link_start {0U};
-  unsigned int link_count {0U};
+  uint64_t link_start {0U};      /* offset into network_v; total edges can exceed 2^32 */
+  unsigned int link_count {0U};  /* per-amplicon, bounded by the variant count */
 };
 
 /* Information about each swarm (cluster) */
@@ -89,7 +89,7 @@ struct Network_state
 {
   std::mutex mutex;
   unsigned int amp {0};
-  unsigned int count {0};
+  uint64_t count {0};  /* running total of network edges; index into network_v */
   std::vector<unsigned int> network_v;
 };
